rename and separate multi-domain files
