--- packages/runner/multi-domain/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "globals": {
-    "window": true
-  }
-}
