--- packages/runner/multi-domain/index.js
@@ -1,1 +1,0 @@
-import '@packages/driver/src/multi-domain'
