+++ packages/runner/src/multi-domain.js
@@ -1,0 +1,2 @@
+// this is the entry point for the multi-domain version of the driver
+import '@packages/driver/src/multi-domain/cypress'
